+++ compiler/oxidate/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod compiler;
+pub mod tests;
